Added a Spelling Corrector feature using python
This Python code allows a user to create a new file or open an existing file and display that text or write new text in the textbox area provided in the user interface. With just a click of the "Correct Spelling" button, all spelling mistakes in the text file that the user opens or creates, can be corrected. The user-friendly interface makes it extremely easy to use.